requirements.txt
Django==1.9.1
djangorestframework==3.3.2
smartconfigparser==0.1.1
django-allauth==0.24.1